requirements.txt
ag2==0.9.6
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
auth0-python==4.10.0
Authlib==1.6.0
autogen-agentchat==0.6.4
autogen-core==0.6.4
autogen-ext==0.6.4
babel==2.17.0
backoff==2.2.1
backrefs==5.9
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
click==8.2.1
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
crewai==0.86.0
crewai-tools==0.25.8
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.9
durationpy==0.10
embedchain==0.1.128
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.13
fastavro==1.11.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
ghp-import==2.1.0
google-auth==2.40.3
google-genai==1.20.0
googleapis-common-protos==1.70.0
gptcache==0.1.44
greenlet==3.2.3
grpcio-health-checking==1.71.2
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.12
idna==3.10
ijson==3.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
instructor==1.9.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
json_repair==0.47.3
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==33.1.0
langsmith==0.3.45
linkup-sdk==0.2.8
litellm==1.72.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.9.4
mcpadapt==0.1.10
mdurl==0.1.2
mem0ai==0.1.111
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.5.0
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
oauthlib==3.3.1
openai==1.95.1
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
paginate==0.5.7
parso==0.8.4
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
portalocker==2.10.1
posthog==3.25.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.6
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.16
pyparsing==3.2.3
pypdf==5.6.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.402
pysbd==0.3.4
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
qdrant-client==1.14.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
schema==0.7.7
scrapegraph_py==1.14.2
seaborn==0.13.2
selenium==4.34.2
serpapi==0.1.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spider-client==0.1.63
SQLAlchemy==2.0.41
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.7.14
uvicorn==0.34.3
uvloop==0.21.0
validators==0.34.0
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
weaviate-client==4.15.4
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
